refactor: entity에서 domain으로 패키지명 변경
