+++ src/main/java/com/example/wini/domain/notification/repository/NotificationRepository.java
@@ -1,7 +1,7 @@
 package com.example.wini.domain.notification.repository;
 
 import com.example.wini.domain.member.domain.Member;
-import com.example.wini.domain.notification.entity.Notification;
+import com.example.wini.domain.notification.domain.Notification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NotificationRepository extends JpaRepository<Notification, Long> {
